backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
cassandra-driver
requests
